[ntuple] simplify TFile anchor class RNTuple
Removes the ROOT::Experimental::Internal::RFileNTupleAnchor class. The
RNTuple anchor class is now standalone (no base class). It moves in its
own compilation unit RNTupleAnchor. This simplifies streaming and schema
evolution.
