+++ tree/ntuple/CMakeLists.txt
@@ -25,6 +25,7 @@ HEADERS
   ROOT/RFieldVisitor.hxx
   ROOT/RMiniFile.hxx
   ROOT/RNTuple.hxx
+  ROOT/RNTupleAnchor.hxx
   ROOT/RNTupleDescriptor.hxx
   ROOT/RNTupleMerger.hxx
   ROOT/RNTupleMetrics.hxx
@@ -51,6 +52,7 @@ SOURCES
   v7/src/REntry.cxx
   v7/src/RMiniFile.cxx
   v7/src/RNTuple.cxx
+  v7/src/RNTupleAnchor.cxx
   v7/src/RNTupleDescriptor.cxx
   v7/src/RNTupleDescriptorFmt.cxx
   v7/src/RNTupleMerger.cxx
