+++ tree/ntuple/inc/LinkDef.h
@@ -14,8 +14,6 @@
 #pragma link off all classes;
 #pragma link off all functions;
 
-// The RNTuple and RFileNTupleAnchor class versions must match
-#pragma link C++ options = version(4) class ROOT::Experimental::Internal::RFileNTupleAnchor + ;
 #pragma link C++ class ROOT::Experimental::RNTuple - ;
 
 #endif
